backend/requirements.txt
fastapi==0.110.0
pydantic==2.7.4
uvicorn==0.18.2
langchain==0.3.16
langchain-community==0.3.16
langchain-core==0.3.32
langchain-text-splitters==0.3.5
pandas==1.5.3
numpy==1.24.4
lz4==4.3.2
yfinance==0.2.52
google-generativeai==0.8.4
beautifulsoup4==4.12.3
